coremind-agent/requirements.txt
langchain
langchain-openai
langchain-community
langchain-ollama 
python-dotenv
pydantic
duckduckgo-search
google-search-results
tiktoken